clipboard-apis/getdata_method_with_wrong_number_of_arguments_paste_on_document.html

<!DOCTYPE html>
<html><head>
    <meta charset="UTF-8">
    <title>Clipboard API - getData() method with wrong number of arguments  - paste on document</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="support/_lib.js"></script>
    <script>
/** getData() method with wrong number of arguments  */
/* Events: paste */
/* paste data: "pasted text" */
function clipboard_api_test(e){
  result(function(){
    assert_throws(new TypeError, function(){
        e.clipboardData.getData();
    });
/*    assert_throws('TypeError', function(){
        e.clipboardData.getData('text/plain', 'wrong number of arguments');
    });*/
  });
}

    </script>
</head>
<body>

    <p>FAILED (This TC requires JavaScript enabled)</p>
    <form>
<input id="input_text" autofocus onfocus="this.select()" value="copied text">
</form>
    <div id="log"></div>
    <script>
    (function(){
        var t = async_test()
        var dataToPaste="pasted text" ;
        var extPassCond=null;
        var eventTarget=document;
        var eventType='paste';
        setup(function(){
                setupTest(t, eventTarget, eventType, dataToPaste, extPassCond );
            });
    })();
    </script>
</body></html>
